requirements.txt
# --- Core App Dependencies ---
streamlit
pandas
geopandas
matplotlib
scikit-learn
python-dotenv
<<<<<<< HEAD
plotly

# --- AI & Language Models ---
openai
langchain
langchain-community
langchain-openai

# --- Azure & Cloud ---
azure-storage-blob
azure-identity
azure-maps-render

# --- Mapping Tools ---
folium
streamlit-folium

# --- BI & Data Export Support ---
pyarrow         # For Parquet (Power BI optimized)
openpyxl        # Write to Excel
xlrd            # Read older Excel formats (optional)

# --- Dev Tools ---
=======
matplotlib
langchain
azure-storage-blob
faiss-cpu
langchain-community
langchain-openai
>>>>>>> de7e8fc3afaa07b13e2567664f434cb0e57d9c62
autopep8
flake8
